backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Bcrypt>=1.0
Flask-JWT-Extended>=4.0
psycopg2-binary>=2.9  # PostgreSQL adapter (use psycopg2 in production)
python-dotenv>=0.19
